SI: add shadow casting, update f32->f64 for higher accuracy during shadow casting

+++ src/camera.rs
@@ -1,19 +1,19 @@
 use crate::{matrix::matrix4d::Matrix4D, transformations::*, point::Point, ray::Ray, world::World, canvas::Canvas};
-use std::{f32, fmt::Error};
+use std::{f64, fmt::Error};
 
 #[derive(Debug)]
 pub struct Camera {
-    pub hsize: f32,
-    pub vsize: f32,
-    pub field_of_view: f32,
+    pub hsize: f64,
+    pub vsize: f64,
+    pub field_of_view: f64,
     pub transform: Matrix4D,
-    pub pixel_size: f32,
-    pub half_width: f32,
-    pub half_height: f32,
+    pub pixel_size: f64,
+    pub half_width: f64,
+    pub half_height: f64,
 }
 
 impl Camera {
-    pub fn new(hsize: f32, vsize: f32, field_of_view: f32) -> Self {
+    pub fn new(hsize: f64, vsize: f64, field_of_view: f64) -> Self {
 
         let half_view = (field_of_view / 2.0).tan();
         let aspect = hsize / vsize;
@@ -39,7 +39,7 @@ impl Camera {
         }
     }
 
-    fn ray_for_pixel(&self, px: f32, py: f32) -> Ray {
+    fn ray_for_pixel(&self, px: f64, py: f64) -> Ray {
 
         let x_offset = (px + 0.5) * self.pixel_size;
         let y_offset = (py + 0.5) * self.pixel_size;
@@ -63,7 +63,7 @@ impl Camera {
 
         for y in 0..self.vsize as usize {
             for x in 0..self.hsize as usize {
-                let r = self.ray_for_pixel(x as f32, y as f32);
+                let r = self.ray_for_pixel(x as f64, y as f64);
                 let color = world.color_at(&r);
                 match color {
                     Ok(color_value) => image.write_pixel(x, y, color_value),
@@ -80,39 +80,39 @@ impl Camera {
 mod tests {
     use crate::{point::Point, vector::Vector, transformations::*, world::World, color::Color};
     use super::Camera;
-    use std::f32;
+    use std::f64;
 
     #[test]
     fn pixel_size_test() {
-        let c = Camera::new(200.0, 125.0, f32::consts::PI/2.0);
+        let c = Camera::new(200.0, 125.0, f64::consts::PI/2.0);
         approx::assert_relative_eq!(c.pixel_size, 0.01);
-        let c = Camera::new(125.0, 200.0, f32::consts::PI/2.0);
+        let c = Camera::new(125.0, 200.0, f64::consts::PI/2.0);
         approx::assert_relative_eq!(c.pixel_size, 0.01);
     }
 
     #[test]
     fn ray_for_pixel_test() {
-        let c = Camera::new(201.0, 101.0, f32::consts::PI/2.0);
+        let c = Camera::new(201.0, 101.0, f64::consts::PI/2.0);
         let r = c.ray_for_pixel(100.0, 50.0);
         approx::assert_relative_eq!(r.origin, Point::new(0.0, 0.0, 0.0));
         approx::assert_relative_eq!(r.direction, Vector::new(0.0, 0.0, -1.0));
 
-        let c = Camera::new(201.0, 101.0, f32::consts::PI/2.0);
+        let c = Camera::new(201.0, 101.0, f64::consts::PI/2.0);
         let r = c.ray_for_pixel(0.0, 0.0);
         approx::assert_relative_eq!(r.origin, Point::new(0.0, 0.0, 0.0));
         approx::assert_relative_eq!(r.direction, Vector::new(0.66519, 0.33259, -0.66851), epsilon=1e-5);
 
-        let mut c = Camera::new(201.0, 101.0, f32::consts::PI/2.0);
-        c.transform = rotation_y(f32::consts::PI/4.0) * translation(0.0, -2.0, 5.0);
+        let mut c = Camera::new(201.0, 101.0, f64::consts::PI/2.0);
+        c.transform = rotation_y(f64::consts::PI/4.0) * translation(0.0, -2.0, 5.0);
         let r = c.ray_for_pixel(100.0, 50.0);
         approx::assert_relative_eq!(r.origin, Point::new(0.0, 2.0, -5.0));
-        approx::assert_relative_eq!(r.direction, Vector::new(2.0f32.sqrt()/2.0, 0.0, -2.0f32.sqrt()/2.0));
+        approx::assert_relative_eq!(r.direction, Vector::new(2.0f64.sqrt()/2.0, 0.0, -2.0f64.sqrt()/2.0));
     }
 
     #[test]
     fn render_test() {
         let w = World::default();
-        let mut c = Camera::new(11.0, 11.0, f32::consts::PI/2.0);
+        let mut c = Camera::new(11.0, 11.0, f64::consts::PI/2.0);
 
         let from = Point::new(0.0, 0.0, -5.0);
         let to = Point::new(0.0, 0.0, 0.0);
